Completed ex 4.21

+++ part4/bloglist-backend/controllers/blogs.js
@@ -51,6 +51,21 @@ blogRouter.post('/', async (request, response, next) => {
 
 blogRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response
+        .status(403)
+        .json({ error: 'no permission to delete this post' })
+        .end()
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     return response.status(204).end()
   } catch (error) {
